feat: Move location address field.
